Move custom nodes from src/evaluation/utils/custom_pipelines.py to deployment/roles/haystack/files/custom_component.py
